Extract array and object print (#9698)
